% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censtreg.R
\name{censtreg}
\alias{censtreg}
\title{Linear regression for data with left-censored response and t-distributed residuals}
\usage{
censtreg(
  formula,
  data,
  censored,
  limit,
  upper = FALSE,
  chains = NULL,
  cores = NULL,
  method = "integrate",
  iter = 2000,
  warmup = 1000,
  lognu_params = c(nu = 6, mu = log(6), sigma = 1),
  sigma_params = c(mu = 1, sigma = 100),
  nu = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data frame. This should contain all the data, including the
censored values. Supposing censoring is from below, setting all response
variable values that are censored to be some value beneath the limit will
result in the correct model. Missing values are not treated as being
censored, but as missing. The function determines what is censored by the
values of the repsonse and the censoring limit, not by missing values or
censoring flags.}

\item{censored}{A string naming a column in \code{data} that indicates if
a response value is censored (\code{censored == 1}) or not (\code{censored == 0}).}

\item{limit}{A vecotor that gives the censoring
point, or a single number giving the censoring point. The former allows
differing censoring points over observations.
Note that if the formula contains a transformation of the response, you
must remember to transform the
limit as well (which is why the argument requires a numeric vector, not the
name of a column in the data). The arguement \code{upper} dictates if the censoring
limit is a lower limit (i.e. left-censoring, censoring from below) or
an upper limit (right-censoring or censoring from above).}

\item{upper}{Logical defaulting to \code{upper = FALSE} indicating that
\code{limit} defines a lower limit, not upper limit, for censoring (i.e.
left-censoring, not right-censoring).}

\item{chains}{The number of Markov chains to run. Defaults to \code{chains = NULL}
and the function guesses the number based on the number of available cores,
specifically, the number of available cores less 1.}

\item{cores}{The number of cores to use. Defaults to \code{cores = NULL} and
the function uses as many cores as chains, or the number of available cores
less 1.}

\item{method}{Either "estimate", in which case the censored values
are treated like parameters and are estimate; or "integrate", in which case
the censored values are integrated out. Defaults to \code{method = "integrate"}.}

\item{iter}{The number of steps to run each Markov chain for. Defaults to
\code{iter = 2000}.}

\item{lognu_params}{Prior parameters for the $t$-distribution used as the prior for
$log nu$. Defaults to \code{lognu_params = c(nu = 6, mu = 6, sigma = 10)}.}

\item{sigma_params}{Prior parameters for the lognormal distribution used as the
prior for $sigma$. Defaults to \code{sigma_params = c(mu = 1, sigma = 100)}.}

\item{nu}{A fixed value of kurtosis parameter nu, which might be helpful if there
are convergence issues and fixing this parameter is acceptable. Putting
\code{nu = Inf} results in the residuals being treated as Gaussian. Defaults
to \code{nu = NULL} and nu is treated as a random parameter in the model.}

\item{silent}{Whether to display Stan's update messages to screen. Defaults
to \code{silent = FALSE}. You'd want to know if the chains were going wrong,
but silencing them can be useful when code is embedded in knitr or
rmarkdown code. If \code{silence = TRUE}, the function tells
\code{rstan::sampling} not to be verbose, not to show warnings, not to
print open progress, and not to refresh. Sheesh!}

\item{...}{Additional arguments to \code{rstan::sampling}.}
}
\description{
Linear regression for data with left-censored response and t-distributed residuals
}
\details{
The function uses the arguments to construct a call to \code{rstan::stan}.
  Not many of the available options for \code{rstan::stan} are manipulable
  through this function, but such things can be easily added if desired.
  If the kurtosis parameter
  is treated as random. Computations are done on the scale of the log of the
  kurtosis parameter, and a t-distribution centered at log(6) with kurtosis parameter
  6 is used as a prior because experience suggests taking logs and giving the
  sampler some clues about likely values aids chain convergence.
  The Stan code is largely copied from Section 4.1, "Censored Data" of the
  Stan User Guide, the differences being the ability to let the location
  depend on covariates, the use of the t-distribution instead of the Gaussian,
  and the prior on the kurtosis parameter.
}
\note{
If the kurtosis parameter is allowed to vary, it is not uncommon for
  some of the Markov chains to exhibit weird behaviour. One solution is to
  specify a tighter prior. Presumably, if you're letting the kurtosis vary,
  you've reason to believe it is somewhere between 1 and 10 at the widest,
  and more likely between 4 and 6, so priors that constrain the majority
  of the mass to those kinds of intervals should only be weakly informative.
}
\seealso{
fitted.censtreg
}
