% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.censtreg}
\alias{fitted.censtreg}
\title{Fitted values from a censtreg object}
\usage{
\method{fitted}{censtreg}(object, method = "lp", what = mean, m = 10, ...)
}
\arguments{
\item{object}{An object of class 'censtreg'.}

\item{method}{One of "lp" for the linear predictor, "summary" for a summary
of the simulated censored values, or "impute" for multiple imputation.}

\item{what}{A function to summarize the coefficients and simulated censored
values. Defaults to \code{what = mean} and the mean is used.}

\item{m}{The number of imputations of the censored response values.}
}
\value{
If the method is "lp" or "summary", a vector; otherwise a matrix.
}
\description{
Fitted values from a censtreg object
}
\details{
It's not very clear to me what use the returned values are when
  \code{method = "summary"}.
}
