% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.censtreg}
\alias{predict.censtreg}
\title{Predicted values from a censtreg object}
\usage{
\method{predict}{censtreg}(
  object,
  newdata,
  se.fit = FALSE,
  ci.fit = TRUE,
  level = 0.95,
  method = "lp",
  what = mean,
  m = 10,
  ...
)
}
\arguments{
\item{object}{An object of class 'censtreg'.}

\item{newdata}{A data frame with the predictors. If not specified, the
function passes its arguments through to \code{fitted} unchanged.}

\item{se.fit}{Whether to include posterior sample deviations of the predictions
in the output. Defaults to \code{se.fit = FALSE}.}

\item{ci.fit}{Whether to include confidence (credible) intervals in the output. Defaults
to \code{ci.fit = TRUE} and percentiles of the posterior predictive
distribution are included.}

\item{level}{The confidence (credible) level when \code{ci.fit = TRUE}.}

\item{method}{One of "lp" for the linear predictor, "summary" for a summary
of the simulated censored values, or "impute" for multiple imputation.
Only used if \code{newdata} is unspecified.}

\item{what}{A function to summarize the coefficients and simulated censored
values. Defaults to \code{what = mean} and the mean is used.
Only used if \code{newdata} is unspecified.}

\item{m}{The number of imputations of the censored response values.
Only used if \code{newdata} is unspecified.}
}
\description{
Predicted values from a censtreg object
}
