% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simCensData}
\alias{simCensData}
\title{Simulate left-censored data to use with \code{censtreg}.}
\usage{
simCensData(n = 1000)
}
\arguments{
\item{n}{The number of datapoints to simulate. Defaults to \code{n = 1000}.}
}
\description{
Simulate left-censored data to use with \code{censtreg}.
}
\details{
A single predictor is simulated, the intercept, slope, scale and
  kurtosis parameters being randomly sampled. The returned object is a list
  containing the data frame and the parameters used to simulate it.
}
