% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{waic.censtreg}
\alias{waic.censtreg}
\title{Compute WAIC from a censtreg object}
\usage{
waic.censtreg(object)
}
\arguments{
\item{object}{A censtreg model, as computed by \code{censtreg}.}
}
\description{
Compute WAIC from a censtreg object
}
\details{
The code has been copied and pasted from the Stan GitHub issues
  pages
}
